feat: refactored valid move predictor + added move prediction for knight, king, rook and bishop
The move predictor now takes in a pieces location and returns a representation of the board with each title storing whether the piece can move to it. This is more efficient and also allows for handling the rook, bishop and queen easily. Prediction for the pieces knight, king, rook and bishop is working but needs to be refactored to remove duplicate code and make it easier to understand. These pieces also need restrictions to prevent them from taking pieces of the players own color.